requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.2

pip~=24.1.2
typing_extensions~=4.12.2
selenium~=4.22.0
requests~=2.32.3
Faker~=26.0.0